Changed how material settings are modified
