+++ Terrain/Core/BaseChunkConfiguration.cs
@@ -8,10 +8,6 @@ namespace GingerVoxelSystem.Core
     [Serializable]
     public class BaseChunkConfiguration : IChunkConfiguration
     {
-        [SerializeField] private DebugOptions debugOptions;
-        public DebugOptions DebugOptions => debugOptions;
-
-
         [SerializeField] private TerrainDensityOptions densityMapOptions;
         public TerrainDensityOptions DensityOptions => densityMapOptions;
