+++ Terrain/Core/Interfaces/IChunkConfiguration.cs
@@ -10,11 +10,6 @@ namespace GingerVoxelSystem.Core
     public interface IChunkConfiguration
     {
         /// <summary>
-        /// Simple debug options mostly for Unity editor.
-        /// </summary>
-        DebugOptions DebugOptions { get; }
-
-        /// <summary>
         /// Density Options used for marching cubes generator.
         /// </summary>
         TerrainDensityOptions DensityOptions { get; }
