+++ Terrain/Systems/Rendering/ChunkRenderRouter.cs
@@ -131,5 +131,4 @@
             }
         }
     }
-
 }
